Make the data more granular
It seems that the retrieval works better when the documents are smaller. I think that this allows it to pull smaller and more relevant sections. I also need to provide facts that can be computed from the data, but can't be derived from the patterns.

Also experimenting with the system prompt to get it to be more truthful.